Проверка доступа
перед выполнением скрипта действия проверяет пользователя (есть ли разрешение на действие)

+++ controller/command/DeleteUserByIdCommand.php
@@ -4,22 +4,28 @@ require_once "./bean/User.php";
 require_once "./service/factory/FactoryService.php";
 require_once "./util/Router.php";
 require_once "./util/permission/PermissionCtrl.php";
+require_once "./resources/conf_const.php";
 
 
-class DeleteUserByIdCommand implements Command
+class DeleteUserByIdCommand extends PermissionCtrl implements Command
 {
+
+    public function __construct()
+    {
+        $this->setAccessedRoles([ADMIN]);
+    }
+
     public function execute(): void
     {
-        PermissionCtrl::permissionCheck('admin');
-
-        $user = $_SESSION['user'];
-        if ($user->getRole() == 'admin') {
-            FactoryService::getInstance()->getUserService()->delete($_POST["user_id_for_deleting"]);
-            header('Location: http://localhost/');
-            exit();
+        if ($this->checkUserPermission()){
+            $user = $_SESSION['user'];
+            if ($user->getRole() == 'admin') {
+                FactoryService::getInstance()->getUserService()->delete($_POST["user_id_for_deleting"]);
+                header('Location: http://localhost/');
+                exit();
+            }
+        }else{
+            Router::redirect();
         }
-        Router::redirect();
-
-
     }
 }
